Fix UCI config read
This commit fixes how values are read from powquty config.
The uci_foreach_element 'loop' left after a continue to look
for a new section, not a new option. This caused some errors if
a option was not set, as all following(in the C code) options
where not read. The same behavior was shown, when an option
value was to long. This is no longer the case and not every
option needs to be set anymore.

Signed-off-by: Stefan Venz <stefan.venz@protonmail.com>